LayoutTests/fast/dom/Window/console-functions-expected.txt

CONSOLE MESSAGE: line 0: console.log() called.
CONSOLE MESSAGE: line 0: console.info() called.
CONSOLE MESSAGE: line 0: console.warn() called.
CONSOLE MESSAGE: line 0: console.error() called.
Tests window.console functions. To make sure this test passed you must check the console for the following 4 messages:

console.log() called.
console.info() called.
console.warn() called.
console.error() called.
